refactor(exception): introduce NotFoundException base class and consolidate 404 handlers
