+++ src/main/java/roomescape/domain/exception/NotFoundException.java
@@ -1,0 +1,7 @@
+package roomescape.domain.exception;
+
+public abstract class NotFoundException extends RuntimeException {
+    protected NotFoundException(String message) {
+        super(message);
+    }
+}
