--- src/main/java/roomescape/service/ReservationNotFoundException.java
@@ -1,7 +1,0 @@
-package roomescape.service;
-
-public class ReservationNotFoundException extends RuntimeException {
-    public ReservationNotFoundException(String message) {
-        super(message);
-    }
-}
