+++ src/main/java/roomescape/service/ReservationTimeService.java
@@ -7,6 +7,7 @@ import org.springframework.transaction.annotation.Transactional;
 import roomescape.dao.ReservationDao;
 import roomescape.dao.ReservationTimeDao;
 import roomescape.domain.ReservationTime;
+import roomescape.service.exception.ReservationConflictException;
 
 @Service
 public class ReservationTimeService {
