+++ src/main/java/roomescape/service/exception/ReservationConflictException.java
@@ -1,4 +1,4 @@
-package roomescape.service;
+package roomescape.service.exception;
 
 public class ReservationConflictException extends RuntimeException {
     public ReservationConflictException(String message) {
