+++ src/main/java/roomescape/service/exception/ReservationNotFoundException.java
@@ -1,0 +1,9 @@
+package roomescape.service.exception;
+
+import roomescape.domain.exception.NotFoundException;
+
+public class ReservationNotFoundException extends NotFoundException {
+    public ReservationNotFoundException(String message) {
+        super(message);
+    }
+}
