+++ src/test/java/roomescape/controller/GlobalExceptionHandlerTest.java
@@ -23,8 +23,8 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
 import roomescape.domain.exception.InvalidInputException;
 import roomescape.domain.exception.PastReservationException;
-import roomescape.service.ReservationConflictException;
-import roomescape.service.ReservationNotFoundException;
+import roomescape.service.exception.ReservationConflictException;
+import roomescape.service.exception.ReservationNotFoundException;
 import roomescape.service.exception.ReservationTimeNotFoundException;
 import roomescape.service.exception.ThemeNotFoundException;
 import roomescape.service.ReservationService;
